pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.11.0.post1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "openai>=1.98.0",
    "openpyxl>=3.1.5",
    "plotly>=6.2.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "streamlit>=1.47.1",
]
